r18322: fixed a compilation problem on AIX caused by lex not putting config.h first. That leads to a conflicting define for lseek() due to _LARGE_FILES being defined after standards headers are included

+++ source/heimdal/lib/asn1/lex.c
@@ -1,3 +1,4 @@
+#include "config.h"
 
 #line 3 "lex.yy.c"
 
@@ -1905,7 +1906,7 @@ static int yy_get_next_buffer (void)
 
 		/* Read in more data. */
 		YY_INPUT( (&YY_CURRENT_BUFFER_LVALUE->yy_ch_buf[number_to_move]),
-			(yy_n_chars), num_to_read );
+			(yy_n_chars), (size_t) num_to_read );
 
 		YY_CURRENT_BUFFER_LVALUE->yy_n_chars = (yy_n_chars);
 		}
